refactor: move resubscription to start events to common class
